requirements.txt
aniso8601==9.0.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Deprecated==1.2.14
envs==1.4
Flask==2.1.3
Flask-Cors==3.0.10
flask-openapi3==2.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
gcloud==0.18.3
googleapis-common-protos==1.60.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauth2client==4.1.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.2
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jwt==4.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.29.0
requests-toolbelt==0.10.1
rpds-py==0.10.3
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
